requirements.txt
pygit2 >= 1.5
pyyaml >= 5.0, < 5.5
click >=8.1, <9.0
jinja2 >= 3.0
deepmerge >=1.0.0
cookiecutter >= 2.0
pydantic >= 1.9